+++ example/src/demos/MultiView.tsx
@@ -5,6 +5,7 @@ import { Canvas, useFrame, useThree, createPortal } from '@react-three/fiber'
 import {
   useGLTF,
   PerspectiveCamera,
+  OrthographicCamera,
   Environment,
   OrbitControls,
   ArcballControls,
@@ -34,7 +35,6 @@ function useHover() {
 }
 
 function View({ index = 1, children, clearColor, placement }: any) {
-  const camera = useRef<THREE.PerspectiveCamera>(null!)
   const { events, size } = useThree()
   const [scene] = useState(() => new THREE.Scene())
   const [position] = useState(() => new THREE.Vector2())
@@ -49,15 +49,15 @@ function View({ index = 1, children, clearColor, placement }: any) {
   useEffect(() => {
     switch (placement) {
       case 'topright':
-        position.set(0.5, 0.5)
+        position.set(1, 1)
         el.style.top = el.style.right = '0px'
         break
       case 'topleft':
-        position.set(0, 0.5)
+        position.set(0, 1)
         el.style.top = el.style.left = '0px'
         break
       case 'bottomright':
-        position.set(0.5, 0)
+        position.set(1, 0)
         el.style.bottom = el.style.right = '0px'
         break
       case 'bottomleft':
@@ -76,22 +76,10 @@ function View({ index = 1, children, clearColor, placement }: any) {
     }
   }, [events, el])
 
-  useFrame((state) => {
-    const left = Math.floor(size.width * position.x)
-    const bottom = Math.floor(size.height * position.y)
-    const width = Math.floor(size.width * 0.5)
-    const height = Math.floor(size.height * 0.5)
-    state.gl.setViewport(left, bottom, width, height)
-    state.gl.setScissor(left, bottom, width, height)
-    state.gl.setScissorTest(true)
-    if (clearColor) state.gl.setClearColor(clearColor)
-    state.gl.render(scene, camera.current)
-  }, index)
-
   const compute = useCallback((event, state) => {
     if (event.target === el) {
-      const width = state.size.width / 2
-      const height = state.size.height / 2
+      const width = state.size.width
+      const height = state.size.height
       state.pointer.set((event.offsetX / width) * 2 - 1, -(event.offsetY / height) * 2 + 1)
       state.raycaster.setFromCamera(state.pointer, state.camera)
     }
@@ -100,32 +88,59 @@ function View({ index = 1, children, clearColor, placement }: any) {
   return (
     <>
       {createPortal(
-        <>
+        <Container index={index} clearColor={clearColor} position={position}>
           {children}
-          <PerspectiveCamera ref={camera} makeDefault />
-        </>,
+        </Container>,
         scene,
-        { events: { compute, priority: events.priority + index, connected: el } },
+        {
+          events: { compute, priority: events.priority + index, connected: el },
+          size: { width: size.width / 2, height: size.height / 2 },
+        },
       )}
     </>
   )
 }
 
+function Container({ children, index, clearColor, position }: any) {
+  const { size, camera, scene } = useThree()
+
+  useFrame((state) => {
+    const left = Math.floor(size.width * position.x)
+    const bottom = Math.floor(size.height * position.y)
+    const width = Math.floor(size.width)
+    const height = Math.floor(size.height)
+    state.gl.setViewport(left, bottom, width, height)
+    state.gl.setScissor(left, bottom, width, height)
+    state.gl.setScissorTest(true)
+    if (clearColor) state.gl.setClearColor(clearColor)
+    state.gl.render(scene, camera)
+  }, index)
+
+  return <>{children}</>
+}
+
 const App = () => (
   <Canvas dpr={[1, 2]} camera={{ position: [0, 0, 5] }}>
     <View index={1} placement="bottomleft" clearColor={new THREE.Color('orange').convertLinearToSRGB()}>
+      <OrthographicCamera makeDefault zoom={100} />
       <Scene preset="lobby" />
+      <OrbitControls makeDefault />
     </View>
     <View index={2} placement="bottomright" clearColor={new THREE.Color('hotpink').convertLinearToSRGB()}>
-      <Scene controls={false} preset="city" />
+      <PerspectiveCamera makeDefault />
+      <Scene preset="city" />
       <ArcballControls />
     </View>
     <View index={3} placement="topleft" clearColor={new THREE.Color('aquamarine').convertLinearToSRGB()}>
+      <PerspectiveCamera makeDefault />
       <Scene preset="dawn" />
+      <OrbitControls makeDefault />
       <CameraShake intensity={2} />
     </View>
     <View index={4} placement="topright" clearColor={new THREE.Color('lightblue').convertLinearToSRGB()}>
+      <PerspectiveCamera makeDefault />
       <Scene preset="warehouse" />
+      <OrbitControls makeDefault />
       <TransformControls scale={3} position={[0, -0.75, 2]}>
         <Soda />
       </TransformControls>
@@ -144,7 +159,6 @@ function Scene({ children, controls = true, preset }: any) {
       <Soda scale={3} position={[0, -0.75, 0]} />
       <Environment preset={preset} />
       {children}
-      {controls && <OrbitControls makeDefault />}
     </Bounds>
   )
 }