+++ packages/fiber/src/native.tsx
@@ -128,13 +128,12 @@ export type {
   Dpr,
   Size,
   Viewport,
-  Camera,
   RenderCallback,
   Performance,
   RootState,
 } from './core/store'
-export type { ThreeEvent, Events, EventManager } from './core/events'
-export type { ObjectMap } from './core/utils'
+export type { ThreeEvent, Events, EventManager, ComputeFunction } from './core/events'
+export type { ObjectMap, Camera } from './core/utils'
 export * from './native/Canvas'
 export { createTouchEvents as events } from './native/events'
 export * from './core'