+++ packages/fiber/src/native/Canvas.tsx
@@ -125,7 +125,9 @@ export const Canvas = /*#__PURE__*/ React.forwardRef<View, Props>(
     }
 
     React.useEffect(() => {
-      return () => unmountComponentAtNode(canvas!)
+      if (canvas) {
+        return () => unmountComponentAtNode(canvas!)
+      }
     }, [canvas])
 
     return (